third_party/WebKit/LayoutTests/editing/spelling/typing_with_composition.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script src="spellcheck_test.js"></script>
<script>
// Regression test for crbug.com/754453
spellcheck_test(
    '<div contenteditable>|</div>',
    document => {
      assert_not_equals(window.textInputController, undefined, 'This test requires textInputController.');
      textInputController.setMarkedText('zz.', 3, 0);
    },
    '<div contenteditable>zz.</div>',
    'Spellcheck should be prevented during text composition.');
</script>
